Separate out CommandOutput related functions
